Add condition to skip pipeline deployment on test
This was implemented because 1/ the pipeline includes stages that require
manual approval, and 2/ the AWS CodeBuild stages build AWS Lambda
functions with elastic network interfaces (ENIs) attached to the
LambdaSecurityGroup resource. As of 2022-02-27, when this type of
CodeBuild job is interrupted, the ENI is orphaned while still attached to
the security group, which will cause stack deletion to fail.